requirements.txt
google-generativeai
python-dotenv
line-bot-sdk
flask
